requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
xgboost>=2.0.0
shap>=0.44.0
imbalanced-learn>=0.12.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
streamlit>=1.31.0
openai>=1.12.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.31.0
